Add quizActionHelpers and refactor startQuiz
